Skip trying to create sequences in Oracle
In Oracle, unlike in Postgres, you have to completely drop and recreate a
sequence if you want to change the current value - or do custom
increment/decrement logic for it. This means that making this work in a
universal way is extremely difficult if not impossible. If this proves to be an
issue in the future, this can be revisited.